Crate document_pii

Modules

item_collector

Contains the helper types and functions which help to iterate over all the given rust files, collect all of the full path names and the actual AST-node for the types defined in those paths. This is later needed for finding PII fields recursively.

pii_finder

This module takes the types and paths from ‘item_collector’ module, and will recursively find all the fields with the specified PII-value, and save those fields with the full path.

Structs

Cli

Prints documentation for metrics.

Output 🔒

Enums

EnumOrStruct

Structs and Enums are the only items that are relevant for finding PII fields.

Functions

find_rs_files 🔒

Gets all the .rs files in a given rust crate/workspace.

main 🔒

print_error 🔒