document_pii::pii_finder

Function get_matching_scoped_paths

fn get_matching_scoped_paths<'a>(
    field_types: &'a BTreeSet<String>,
    scoped_paths: &'a BTreeSet<String>,
) -> Vec<&'a String>

Finds the full path to the given types by comparing them to the types in the scope.