Struct document_pii::pii_finder::PiiFinder

pub struct PiiFinder<'a> {
    pub module_path: String,
    pub current_type: String,
    pub all_types: &'a HashMap<String, EnumOrStruct>,
    pub scoped_paths: &'a BTreeMap<String, BTreeSet<String>>,
    pub current_path: Vec<TypeAndField>,
    pub pii_types: BTreeSet<FieldsWithAttribute>,
}

This is the visitor that actually generates the pii_types, it has a lot of associated data because it using the Visit trait from syn-crate means I cannot add data as arguments. The ‘pii_types’ field can be regarded as the output.

Fields
module_path: String

Module path of a type is the full path up to the type itself.

Example: relay_event_schema::protocol::Event -> relay_event_schema::protocol

current_type: Stringall_types: &'a HashMap<String, EnumOrStruct>scoped_paths: &'a BTreeMap<String, BTreeSet<String>>current_path: Vec<TypeAndField>pii_types: BTreeSet<FieldsWithAttribute>
Implementations

impl<'a> PiiFinder<'a>

pub fn new( path: &str, all_types: &'a HashMap<String, EnumOrStruct>, scoped_paths: &'a BTreeMap<String, BTreeSet<String>> ) -> Result<Self>

fn visit_type_path(&mut self, path: &TypePath)

fn visit_field_types(&mut self, ty: &Type)

fn is_current_type_already_visited(&self) -> bool

Checks if the type we are on has already been visited, this is to avoid infinite recursion.