Module document_pii::item_collector

Contains the helper types and functions which help to iterate over all the given rust files, collect all of the full path names and the actual AST-node for the types defined in those paths. This is later needed for finding PII fields recursively.

Structs

AstItemCollector

The types and use statements items collected from the rust files.

TypesAndScopedPaths

Functions

add_file_stem_to_module_path 🔒

crate_name_from_file 🔒

flatten_use_tree 🔒

Flattens a usetree.

is_file_declared_from_mod_file 🔒

is_file_declared_from_other_file 🔒

is_file_module 🔒

module_name_from_file 🔒

Takes in the path to a Rust file and returns the path as you’d refer to it in a use-statement.

normalize_type_path 🔒

usetree_to_paths 🔒

First flattens the UseTree and then normalizing the paths.